rename: LadderController -> LadderGameApplication
