+++ src/main/java/Application.java
@@ -1,4 +1,4 @@
-import controller.LadderController;
+import application.LadderGameApplication;
 import utils.RandomNumberGenerator;
 import view.InputView;
 import view.OutputView;
@@ -6,8 +6,8 @@ import view.OutputView;
 public class Application {
 
     public static void main(String[] args) {
-        LadderController ladderController = new LadderController(new InputView(), new OutputView(),
-                new RandomNumberGenerator());
-        ladderController.run();
+        LadderGameApplication ladderGameApp = new LadderGameApplication(
+                new InputView(), new OutputView(), new RandomNumberGenerator());
+        ladderGameApp.run();
     }
 }
